Extracted common functions
