+++ src/pages/api/events.ts
@@ -1,75 +1,27 @@
 import { NextApiRequest, NextApiResponse } from "next"
 
+import {
+    checkRequestHeader,
+    ContentType,
+    greet,
+    PackedRequestData,
+    printRequest,
+    respond,
+} from "../../libs/apiHandling"
 import connect from "../../libs/database"
 
 import Event, { eventSchemaKeys } from "../../model/event"
-
-interface PackedRequestData {
-    client: string
-    req: NextApiRequest
-    res: NextApiResponse
-}
-
-/** HTTP Response Code-String Pair */
-interface HttpResponseCodeStringPair {
-    [key: number]: string
-}
-
-/** HTTP Response Code */
-const HRC: HttpResponseCodeStringPair = {
-    200: "200 OK",
-    201: "201 Created",
-    400: "400 Bad Request",
-    404: "404 Not Found",
-    405: "405 Method Not Allowed",
-    409: "409 Conflict",
-    415: "415 Unsupported Media Type",
-}
-
-const enum ContentType {
-    JSON = "application/json",
-    PLAINTEXT = "text/plain",
-}
-
-function checkRequestHeader(r: PackedRequestData, contentType: ContentType) {
-    const { client, req } = r
-
-    const requestContentType = req.headers["content-type"]
-    if (requestContentType !== contentType) {
-        console.log(
-            `[${client}] sent "${requestContentType}", but we are expecting Content-Type "${contentType}"`
-        )
-        respond(r, 415)
-        return false
-    }
-
-    return true
-}
-
-function printRequest(r: PackedRequestData) {
-    const { client, req } = r
-
-    console.log(`[${client}]'s Request Header:\n\n${JSON.stringify(req.headers, null, 4)}\n`)
-    console.log(`[${client}]'s Request Body:\n\n${JSON.stringify(req.body, undefined, 4)}\n`)
-}
-
-function respond(r: PackedRequestData, code: number) {
-    const { client, res } = r
-
-    console.log(`[${client}] is to be responded with: ${HRC[code]}`)
-    res.status(code).end()
-}
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     await connect()
 
     const client = `${req.socket.remoteAddress}:${req.socket.remotePort}`
-    console.log(`[${client}] says ${req.method} to ${req.url}`)
-
     const r: PackedRequestData = { client, req, res }
     const resp = (code: number) => {
         respond(r, code)
     }
+
+    greet(r)
 
     switch (req.method) {
         case "POST":
